% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnAge_variables.R
\name{AnAge_variables}
\alias{AnAge_variables}
\title{Potential variables to extract from the AnAge database}
\description{
Potential variables to extract from the AnAge database
}
\references{
Tacutu, R., Thornton, D., Johnson, E., Budovsky, A., Barardo, D., Craig, T., Diana, E.,
Lehmann, G., Toren, D., Wang, J., Fraifeld, V. E., de Magalhaes, J. P. (2018)
\emph{Human Ageing Genomic Resources: new and updated databases.}. Nucleic Acids Research 46(D1):D1083-D1090.
}
\keyword{datasets}
