% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_AnAge.R
\name{scrape_AnAge}
\alias{scrape_AnAge}
\title{Extracting life history data from the AnAge database}
\usage{
scrape_AnAge(latin_name = NULL, vars = NULL, name_sep = " ",
  download_data = TRUE, file_path = "life_history_data.zip")
}
\arguments{
\item{latin_name}{latin name of the species. Can be a string "Felis catus" or a vector with
multiple species c("Felis catus", "Bos taurus"). The default separator is " ",
but others can be chosen, see below. Should be case insensitive.}

\item{vars}{vector of variables to extract for the species. Example:
 c("maximum_lovevity_yrs", "female_matury_days"). \cr
 Options are: \cr
"maximum_longevity_yrs" \cr
"female_maturity_days"  \cr
"male_maturity_days"    \cr
"gestation_incubation_days" - duration of either gestation or incubation in days \cr
"weaning_days" \cr
"litter_clutch_size" - size of either litter or clutch \cr
\cr
To see all options, check data(AnAge_variables)}

\item{name_sep}{separator between genus and species in latin_name. Default is " ", as in "Felis catus".}

\item{download_data}{Data has to be downloaded at least once and is by default
saved into the working directory. For each subsequent call, download_data can be set to FALSE,
except for when you want to update the datafile.}

\item{file_path}{Name and path of file. Make sure to keep it as a .zip. Defaults to
"life_history_data.zip", thus saving it into the working directory or looking
for it in the working directory when download_data = FALSE.}
}
\description{
An R wrapper to access life history data from AnAge (http://genomics.senescence.info/species/)
for one or multiple species.
}
\examples{

dat <- scrape_AnAge(latin_name = c("felis catus"),
                    vars = c("maximum_longevity_yrs","female_maturity_days"))
dat

# For multiple species. Now download_data = FALSE, as data has been
# downloaded already in the example above.

species <- c("felis catus", "Sterna paradisaea")
dat <- scrape_AnAge(latin_name = species,
                    vars = c("maximum_longevity_yrs","female_maturity_days"),
                    download_data = FALSE)
dat

}
\references{
Tacutu, R., Thornton, D., Johnson, E., Budovsky, A., Barardo, D., Craig, T., Diana, E.,
Lehmann, G., Toren, D., Wang, J., Fraifeld, V. E., de Magalhaes, J. P. (2018)
\emph{Human Ageing Genomic Resources: new and updated databases.}. Nucleic Acids Research 46(D1):D1083-D1090.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com)
}
